Introducing http filter chain
Quicsec now provide means to make the requests to pass throught a filter chain.
When any filter has its execution interreput by the policies the request is denied. The filter chain can accept
any amount of concatened filters or none. Initialiy we are provinding to filters: CorazaFilter which utilizes
coraza WAF and also the ExtAuthFFilter that uses OPA.